Fixed routing bug and added file filters
There was a routing bug where a . in a
route would confuse the config class into
thinking it was an array. Now config files
are included without you specifying the
extension, and config files with the
extension "strconf" will be interpreted
without any regard to periods.

File filters allow you to send files through
a series of filters before using them.
A filter could be a css minification script
or a template parsing engine. So far
only filtered views can be cached. Filters
are specified as follows: something.min.css
will be passed through the "min" filter.
other.tpl.min.css will be passed through
the tpl filter and then the min filter.
The min filter minifies css and the tpl
filter will be a template parser.